update feature: implement ui
+ pause state panel
