+++ Assets/Font/Bungee-Regular SDF.asset
@@ -60,9 +60,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 71c1514a6bd24e1e882cebbe1904ce04, type: 3}
   m_Name: Bungee-Regular SDF
   m_EditorClassIdentifier: 
-  hashCode: 0
+  hashCode: -1033442216
   material: {fileID: 3094482421874300036}
-  materialHashCode: 0
+  materialHashCode: 246761848
   m_Version: 1.1.0
   m_SourceFontFileGUID: db6b308f3a830459d89e9fcce001314b
   m_SourceFontFile_EditorRef: {fileID: 12800000, guid: db6b308f3a830459d89e9fcce001314b, type: 3}
@@ -2663,9 +2663,9 @@ Material:
     - _OutlineWidth: 0
     - _PerspectiveFilter: 0.875
     - _Reflectivity: 10
-    - _ScaleRatioA: 1
-    - _ScaleRatioB: 1
-    - _ScaleRatioC: 1
+    - _ScaleRatioA: 0.8333333
+    - _ScaleRatioB: 0.6770833
+    - _ScaleRatioC: 0.6770833
     - _ScaleX: 1
     - _ScaleY: 1
     - _ShaderFlags: 0
