+++ Assets/Font/Lato-Regular SDF.asset
@@ -73,9 +73,9 @@ Material:
     - _OutlineWidth: 0
     - _PerspectiveFilter: 0.875
     - _Reflectivity: 10
-    - _ScaleRatioA: 1
-    - _ScaleRatioB: 1
-    - _ScaleRatioC: 1
+    - _ScaleRatioA: 0.8333333
+    - _ScaleRatioB: 0.6770833
+    - _ScaleRatioC: 0.6770833
     - _ScaleX: 1
     - _ScaleY: 1
     - _ShaderFlags: 0
@@ -120,9 +120,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 71c1514a6bd24e1e882cebbe1904ce04, type: 3}
   m_Name: Lato-Regular SDF
   m_EditorClassIdentifier: 
-  hashCode: 0
+  hashCode: 1753386352
   material: {fileID: -421686805889537028}
-  materialHashCode: 0
+  materialHashCode: -1935499824
   m_Version: 1.1.0
   m_SourceFontFileGUID: 691eb6d702c804235a026763b6795515
   m_SourceFontFile_EditorRef: {fileID: 12800000, guid: 691eb6d702c804235a026763b6795515, type: 3}
