+++ Assets/Minigame 3/UI/Pause.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 40327f50e979f9e4aa164456beb8a4cb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
